+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change history for ui-plugin-query-builder
 
+* [UIPQB-157](https://folio-org.atlassian.net/browse/UIPQB-157) Add stop polling mechanism for useAsyncDataSource after 3 retries.
+
 ## [1.2.3](https://github.com/folio-org/ui-plugin-query-builder/tree/v1.2.3) (2024-11-12)
 
 * [UIPQB-126](https://folio-org.atlassian.net/browse/UIPQB-126) Convert local date to UTC with respect to tenant zone, rather than user.